Move AccessSolver implementation to postgres-core-resolver (from postgres-graphql-resolver)
