+++ crates/postgres-subsystem/postgres-core-resolver/Cargo.toml
@@ -14,6 +14,7 @@ network = [
 bigdecimal = ["exo-sql/bigdecimal"]
 
 [dependencies]
+async-trait.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 tokio-postgres.workspace = true
@@ -21,10 +22,14 @@ base64 = "0.22.1"
 chrono.workspace = true
 serde_json.workspace = true
 uuid.workspace = true
+maybe-owned.workspace = true
+futures.workspace = true
+
 
 exo-sql = { path = "../../../libs/exo-sql" }
 exo-env = { path = "../../../libs/exo-env" }
 common = { path = "../../common" }
+postgres-core-model = { path = "../postgres-core-model" }
 core-plugin-interface = { path = "../../core-subsystem/core-plugin-interface" }
 
 [dev-dependencies]
