+++ crates/postgres-subsystem/postgres-core-resolver/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod access;
 pub mod cast;
 pub mod database_helper;
 pub mod postgres_execution_error;
