+++ crates/postgres-subsystem/postgres-graphql-resolver/Cargo.toml
@@ -14,7 +14,8 @@ futures.workspace = true
 indexmap.workspace = true
 serde_json = { workspace = true, features = ["preserve_order"] }
 tokio.workspace = true
-maybe-owned.workspace = true
+http.workspace = true
+
 
 exo-sql = { path = "../../../libs/exo-sql" }
 exo-env = { path = "../../../libs/exo-env" }
