+++ crates/postgres-subsystem/postgres-graphql-resolver/src/access/mod.rs
@@ -1,6 +1,5 @@
-pub mod access_op;
-pub mod database_solver;
-pub mod precheck_solver;
+mod database_solver;
+mod precheck_solver;
 
 #[cfg(test)]
 mod test_util;
