+++ crates/postgres-subsystem/postgres-graphql-resolver/src/access/precheck_solver/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(test)]
+mod tests;
+
+#[cfg(test)]
+mod test_system;
